+++ src/helpers/repository.ts
@@ -14,11 +14,13 @@ export const WEB_IDES = [
     title: "github.dev",
     baseUrl: "https://github.dev/",
     icon: { source: "github-dev.svg", tintColor: Color.PrimaryText },
+    shortcut: { modifiers: ["cmd"], key: "o" },
   },
   {
     title: "VS Code for the Web",
     baseUrl: "https://vscode.dev/github/",
     icon: { source: "vscode.svg", tintColor: Color.PrimaryText },
+    shortcut: { modifiers: ["cmd"], key: "v" },
   },
   {
     title: "CodeSandbox",
@@ -44,6 +46,13 @@ export const WEB_IDES = [
     title: "Sourcegraph",
     baseUrl: `https://sourcegraph.com/github.com/`,
     icon: { source: "sourcegraph.svg", tintColor: Color.PrimaryText },
+    shortcut: { modifiers: ["cmd"], key: "s" },
+  },
+  {
+    title: "DeepWiki",
+    baseUrl: "https://deepwiki.com/",
+    icon: { source: "deepwiki.png" },
+    shortcut: { modifiers: ["cmd"], key: "d" },
   },
   {
     title: "VS Code Remote Repositories",